+++ test/data/test_multi_nested_tensor.py
@@ -1,8 +1,21 @@
 import random
+from typing import List
 
+import pytest
 import torch
+from torch import Tensor
 
 from torch_frame.data import MultiNestedTensor
+
+
+def assert_equal(tensor_mat: List[List[Tensor]],
+                 multi_nested_tensor: MultiNestedTensor):
+    assert len(tensor_mat) == multi_nested_tensor.shape[0]
+    assert len(tensor_mat[0]) == multi_nested_tensor.shape[1]
+    for i in range(multi_nested_tensor.shape[0]):
+        for j in range(multi_nested_tensor.shape[1]):
+            tensor = multi_nested_tensor[i, j]
+            assert torch.allclose(tensor_mat[i][j], tensor)
 
 
 def test_multi_nested_tensor_basic():
@@ -19,9 +32,17 @@ def test_multi_nested_tensor_basic():
 
     multi_nested_tensor = MultiNestedTensor.from_tensor_mat(tensor_mat)
     assert str(multi_nested_tensor
-               ) == "MultiNestedTensor(num_rows=8, num_cols=10, device=cpu)"
-    assert multi_nested_tensor.num_rows == num_rows
-    assert multi_nested_tensor.num_cols == num_cols
+               ) == "MultiNestedTensor(num_rows=8, num_cols=10, device='cpu')"
+
+    # Test sizes
+    assert multi_nested_tensor.shape[0] == num_rows
+    assert multi_nested_tensor.size(0) == num_rows
+    assert multi_nested_tensor.shape[1] == num_cols
+    assert multi_nested_tensor.size(1) == num_cols
+    with pytest.raises(ValueError, match="not have a fixed length"):
+        multi_nested_tensor.size(2)
+    with pytest.raises(IndexError, match="Dimension out of range"):
+        multi_nested_tensor.size(3)
 
     # Test multi_nested_tensor[i, j] indexing
     for i in range(-num_rows, num_rows):
@@ -33,26 +54,83 @@ def test_multi_nested_tensor_basic():
     # Test multi_nested_tensor[i] indexing
     for i in range(-num_rows, num_rows):
         multi_nested_tensor_row = multi_nested_tensor[i]
-        assert multi_nested_tensor_row.num_rows == 1
-        assert multi_nested_tensor_row.num_cols == num_cols
+        assert multi_nested_tensor_row.shape[0] == 1
+        assert multi_nested_tensor_row.shape[1] == num_cols
         for j in range(-num_cols, num_cols):
             tensor = multi_nested_tensor_row[0, j]
             assert isinstance(tensor, torch.Tensor)
             assert torch.allclose(tensor_mat[i][j], tensor)
 
+    # Test row slicing
+    assert_equal(tensor_mat, multi_nested_tensor[:])
+    assert_equal(tensor_mat[:3], multi_nested_tensor[:3])
+    assert_equal(tensor_mat[3:], multi_nested_tensor[3:])
+    assert_equal(tensor_mat[3:5], multi_nested_tensor[3:5])
+    assert_equal(tensor_mat[-7:5], multi_nested_tensor[-7:5])
+    assert_equal(tensor_mat[-7:-1], multi_nested_tensor[-7:-1])
+    assert_equal(tensor_mat[1::2], multi_nested_tensor[1::2])
+    empty_multi_nested_tensor = multi_nested_tensor[5:3]
+    assert empty_multi_nested_tensor.shape[0] == 0
+    assert empty_multi_nested_tensor.shape[1] == num_cols
+
+    # Test row narrow
+    assert_equal(tensor_mat[3:3 + 2],
+                 multi_nested_tensor.narrow(dim=0, start=3, length=2))
+
     # Test multi_nested_tensor[List[int]] indexing
     for index in [[4], [2, 2], [-4, 1, 7], [3, -7, 1, 0]]:
         multi_nested_tensor_indexed = multi_nested_tensor[index]
-        assert multi_nested_tensor_indexed.num_rows == len(index)
-        assert multi_nested_tensor_indexed.num_cols == num_cols
+        assert multi_nested_tensor_indexed.shape[0] == len(index)
+        assert multi_nested_tensor_indexed.shape[1] == num_cols
         for i, idx in enumerate(index):
             for j in range(num_cols):
                 tensor = multi_nested_tensor_indexed[i, j]
                 assert torch.allclose(tensor_mat[idx][j], tensor)
 
-    cloned_multi_nested_tensor = multi_nested_tensor.clone()
+    # Test multi_nested_tensor[:, i] indexing
+    for j in range(-num_cols, num_cols):
+        multi_nested_tensor_col = multi_nested_tensor[:, j]
+        assert multi_nested_tensor_col.shape[0] == num_rows
+        assert multi_nested_tensor_col.shape[1] == 1
+        for i in range(-num_rows, num_rows):
+            tensor = multi_nested_tensor_col[i, 0]
+            assert isinstance(tensor, torch.Tensor)
+            assert torch.allclose(tensor_mat[i][j], tensor)
 
+    # Test column List[int] indexing
+    with pytest.raises(NotImplementedError):
+        # TODO: Add proper test once implemented
+        multi_nested_tensor[:, [2, 4, 6]]
+
+    # Test column slicing
+    with pytest.raises(NotImplementedError):
+        # TODO: Add proper test once implemented
+        multi_nested_tensor[:, 4:8]
+
+    # Test column narrow
+    with pytest.raises(NotImplementedError):
+        # TODO: Add proper test once implemented
+        multi_nested_tensor.narrow(dim=1, start=3, length=2)
+
+    # Testing clone
+    cloned_multi_nested_tensor = multi_nested_tensor.clone()
     multi_nested_tensor.values[0] = max_value + 1.0
     assert cloned_multi_nested_tensor.values[0] != max_value + 1.0
     multi_nested_tensor.offset[0] = -1
     assert cloned_multi_nested_tensor.values[0] != -1
+
+
+def test_multi_nested_tensor_different_num_rows():
+    tensor_mat = [
+        [torch.tensor([1, 2, 3]),
+         torch.tensor([4, 5])],
+        [torch.tensor([6, 7]),
+         torch.tensor([8, 9, 10]),
+         torch.tensor([11])],
+    ]
+
+    with pytest.raises(
+            RuntimeError,
+            match="The length of each row must be the same",
+    ):
+        MultiNestedTensor.from_tensor_mat(tensor_mat)